test(vla): establish official upstream parity
Pin the LingBot-VLA v2 reference checkout and add an isolated upstream capture runtime for preprocessing, velocity, and final-action comparison.

Commit the official RobotWin normalization statistics, align public image preprocessing and normalization precision, and record attention, MoE, and norm-stat identities in parity artifacts. Keep the upstream Triton MoE path enabled for production while providing a deterministic reference backend for strict cross-process comparison.

Verification: .venv-vla/bin/python -m pytest tests/unit/models/test_lingbot_vla_v2_loader.py tests/unit/pipelines/lingbot_vla_v2/test_data.py tests/unit/pipelines/lingbot_vla_v2/test_robot_profile.py tests/unit/validation/test_lingbot_vla_v2_artifacts.py -q (24 passed); ruff check and ruff format --check passed for the changed VLA pipeline, tests, and validation scripts; git diff --cached --check passed.

+++ telefuser/pipelines/lingbot_vla_v2/data.py
@@ -94,9 +94,7 @@ class LingBotVlaV2InputProcessor:
                 f"model max_state_dim is {self.max_state_dim}, RobotWin requires {robot_profile.canonical_dim}"
             )
 
-    def _process_images(
-        self, images: Mapping[str, ImageInput]
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+    def _process_images(self, images: Mapping[str, ImageInput]) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
         missing = [key for key in ROBOTWIN_CAMERA_KEYS if key not in images]
         if missing:
             raise ValueError(f"RobotWin observation is missing camera keys: {missing}")
@@ -104,7 +102,7 @@ class LingBotVlaV2InputProcessor:
         processed_images: list[torch.Tensor] = []
         grids: list[torch.Tensor] = []
         for key in self.robot_profile.camera_keys:
-            image = self.image_resize(_image_to_chw_uint8(images[key]))
+            image = self.image_resize(_image_to_chw_uint8(images[key]).to(dtype=torch.float32))
             output = self.processor.image_processor(image)
             pixels = output["pixel_values"] if isinstance(output, dict) else output.pixel_values
             grid = output.get("image_grid_thw") if isinstance(output, dict) else getattr(output, "image_grid_thw", None)
@@ -113,9 +111,7 @@ class LingBotVlaV2InputProcessor:
             if pixels.ndim == 3 and pixels.shape[0] == 1:
                 pixels = pixels.squeeze(0)
             if pixels.ndim != 2:
-                raise ValueError(
-                    f"Qwen3-VL image processor must return [patches, features], got {tuple(pixels.shape)}"
-                )
+                raise ValueError(f"Qwen3-VL image processor must return [patches, features], got {tuple(pixels.shape)}")
             if grid is None or grid.numel() < 3:
                 raise ValueError("Qwen3-VL image processor must return image_grid_thw")
             processed_images.append(pixels)
